fix: 바텀시트 read type 추가 변경
